feat : 서비스 인터페이스 구현
