+++ src/main/java/ddingdong/ddingdongBE/domain/documents/controller/AdminDocumentController.java
@@ -4,7 +4,7 @@ import ddingdong.ddingdongBE.auth.PrincipalDetails;
 import ddingdong.ddingdongBE.domain.documents.api.AdminDocumentApi;
 import ddingdong.ddingdongBE.domain.documents.controller.dto.request.CreateDocumentRequest;
 import ddingdong.ddingdongBE.domain.documents.controller.dto.request.UpdateDocumentRequest;
-import ddingdong.ddingdongBE.domain.documents.service.FacadeAdminDocumentService;
+import ddingdong.ddingdongBE.domain.documents.service.FacadeAdminDocumentServiceImpl;
 import ddingdong.ddingdongBE.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -15,24 +15,24 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class AdminDocumentController implements AdminDocumentApi {
 
-    private final FacadeAdminDocumentService facadeAdminDocumentService;
+    private final FacadeAdminDocumentServiceImpl facadeAdminDocumentServiceImpl;
 
     public void createDocument(
         @AuthenticationPrincipal PrincipalDetails principalDetails,
         CreateDocumentRequest createDocumentRequest
     ) {
         User admin = principalDetails.getUser();
-        facadeAdminDocumentService.create(createDocumentRequest.toCommand(admin));
+        facadeAdminDocumentServiceImpl.create(createDocumentRequest.toCommand(admin));
     }
 
     public void updateDocument(
         @PathVariable Long documentId,
         UpdateDocumentRequest updateDocumentRequest
     ) {
-        facadeAdminDocumentService.update(updateDocumentRequest.toCommand(documentId));
+        facadeAdminDocumentServiceImpl.update(updateDocumentRequest.toCommand(documentId));
     }
 
     public void deleteDocument(@PathVariable Long documentId) {
-        facadeAdminDocumentService.delete(documentId);
+        facadeAdminDocumentServiceImpl.delete(documentId);
     }
 }
