+++ src/main/java/ddingdong/ddingdongBE/domain/documents/controller/DocumentController.java
@@ -4,7 +4,7 @@ import ddingdong.ddingdongBE.domain.documents.api.DocumentApi;
 import ddingdong.ddingdongBE.domain.documents.controller.dto.request.GetDocumentPagingRequest;
 import ddingdong.ddingdongBE.domain.documents.controller.dto.response.DocumentListResponse;
 import ddingdong.ddingdongBE.domain.documents.controller.dto.response.DocumentResponse;
-import ddingdong.ddingdongBE.domain.documents.service.FacadeDocumentService;
+import ddingdong.ddingdongBE.domain.documents.service.FacadeDocumentServiceImpl;
 import ddingdong.ddingdongBE.domain.documents.service.dto.query.DocumentListPagingQuery;
 import ddingdong.ddingdongBE.domain.documents.service.dto.query.DocumentQuery;
 import lombok.RequiredArgsConstructor;
@@ -15,17 +15,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class DocumentController implements DocumentApi {
 
-    private final FacadeDocumentService facadeDocumentService;
+    private final FacadeDocumentServiceImpl facadeDocumentServiceImpl;
 
     @Override
     public DocumentListResponse getDocumentList(GetDocumentPagingRequest request) {
-        DocumentListPagingQuery query = facadeDocumentService.getDocumentList(request.toCommand());
+        DocumentListPagingQuery query = facadeDocumentServiceImpl.getDocumentList(request.toCommand());
         return DocumentListResponse.from(query);
     }
 
     @Override
     public DocumentResponse getDocument(@PathVariable Long documentId) {
-        DocumentQuery query = facadeDocumentService.getDocument(documentId);
+        DocumentQuery query = facadeDocumentServiceImpl.getDocument(documentId);
         return DocumentResponse.from(query);
     }
 }
