+++ src/test/java/ddingdong/ddingdongBE/common/support/WebApiUnitTestSupport.java
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import ddingdong.ddingdongBE.domain.club.service.ClubService;
 import ddingdong.ddingdongBE.domain.documents.controller.AdminDocumentController;
 import ddingdong.ddingdongBE.domain.documents.controller.DocumentController;
-import ddingdong.ddingdongBE.domain.documents.service.FacadeAdminDocumentService;
-import ddingdong.ddingdongBE.domain.documents.service.FacadeDocumentService;
+import ddingdong.ddingdongBE.domain.documents.service.FacadeAdminDocumentServiceImpl;
+import ddingdong.ddingdongBE.domain.documents.service.FacadeDocumentServiceImpl;
 import ddingdong.ddingdongBE.domain.fileinformation.service.FileInformationService;
 import ddingdong.ddingdongBE.file.service.FileService;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,9 +32,9 @@ public abstract class WebApiUnitTestSupport {
   @Autowired
   protected MockMvc mockMvc;
   @MockBean
-  protected FacadeDocumentService facadeDocumentService;
+  protected FacadeDocumentServiceImpl facadeDocumentServiceImpl;
   @MockBean
-  protected FacadeAdminDocumentService facadeAdminDocumentService;
+  protected FacadeAdminDocumentServiceImpl facadeAdminDocumentServiceImpl;
   @MockBean
   protected FileService fileService;
   @MockBean
