+++ src/test/java/ddingdong/ddingdongBE/domain/documents/controller/AdminDocumentControllerUnitTest.java
@@ -31,7 +31,7 @@ public class AdminDocumentControllerUnitTest extends WebApiUnitTestSupport {
             .title("새로운 문서 제목")
             .fileIds(List.of("1","2"))
             .build();
-        doNothing().when(facadeAdminDocumentService).create(any(CreateDocumentCommand.class));
+        doNothing().when(facadeAdminDocumentServiceImpl).create(any(CreateDocumentCommand.class));
 
         // when // then
         mockMvc.perform(MockMvcRequestBuilders.post("/server/admin/documents")
@@ -41,7 +41,7 @@ public class AdminDocumentControllerUnitTest extends WebApiUnitTestSupport {
             .andDo(print())
             .andExpect(status().isCreated());
 
-        verify(facadeAdminDocumentService).create(any());
+        verify(facadeAdminDocumentServiceImpl).create(any());
     }
 
     @WithMockAuthenticatedUser(role = "ADMIN")
@@ -56,7 +56,7 @@ public class AdminDocumentControllerUnitTest extends WebApiUnitTestSupport {
             .fileIds(List.of("1","2"))
             .build();
 
-        doNothing().when(facadeAdminDocumentService).update(any(UpdateDocumentCommand.class));
+        doNothing().when(facadeAdminDocumentServiceImpl).update(any(UpdateDocumentCommand.class));
 
         // when // then
         mockMvc.perform(MockMvcRequestBuilders.patch("/server/admin/documents/{documentId}", updateId)
@@ -65,7 +65,7 @@ public class AdminDocumentControllerUnitTest extends WebApiUnitTestSupport {
                 .content(objectMapper.writeValueAsString(request)))
                 .andExpect(status().isOk());
 
-        verify(facadeAdminDocumentService).update(any());
+        verify(facadeAdminDocumentServiceImpl).update(any());
     }
 
     @WithMockAuthenticatedUser(role = "ADMIN")
@@ -80,6 +80,6 @@ public class AdminDocumentControllerUnitTest extends WebApiUnitTestSupport {
                 .andDo(print())
                 .andExpect(status().isOk());
 
-        verify(facadeAdminDocumentService).delete(deletedId);
+        verify(facadeAdminDocumentServiceImpl).delete(deletedId);
     }
 }
